requirements.txt
streamlit
pyrebase4
python-dotenv
pandas
matplotlib
pymupdf
openai
requests
